Fix casts.

+++ src/include/hardware.h
@@ -28,11 +28,11 @@
 /*
  * Atomic versions of the flag set/clear macros.
  */
-#define F_ISSET_ATOMIC(p, mask) ((p)->flags_atomic & (uint8_t)(mask))
+#define F_ISSET_ATOMIC(p, mask) ((p)->flags_atomic & (uint16_t)(mask))
 
 #define F_SET_ATOMIC(p, mask)                                                                \
     do {                                                                                     \
-        uint8_t __orig;                                                                      \
+        uint16_t __orig;                                                                     \
         do {                                                                                 \
             __orig = (p)->flags_atomic;                                                      \
         } while (!__wt_atomic_cas16(&(p)->flags_atomic, __orig, __orig | (uint16_t)(mask))); \
@@ -40,7 +40,7 @@
 
 #define F_CLR_ATOMIC(p, mask)                                                                 \
     do {                                                                                      \
-        uint8_t __orig;                                                                       \
+        uint16_t __orig;                                                                      \
         do {                                                                                  \
             __orig = (p)->flags_atomic;                                                       \
         } while (!__wt_atomic_cas16(&(p)->flags_atomic, __orig, __orig & ~(uint16_t)(mask))); \
